update rake register

+++ lib/tasks/register.rake
@@ -15,6 +15,7 @@ namespace :register do
         store(Area, 'area.csv') do |row|
           Area.new(row.to_hash.merge("polygon"=>nil)).save!
         end
+        store(Agent, 'agent.csv')
         store(PredefinedPosition, 'predefined_position.csv')
         store(GatheringPosition, 'gathering_position.csv')
         store(Department, 'department.csv')
@@ -27,17 +28,19 @@ namespace :register do
   # ※ブロック指定時はブロックに処理を提供する
   def store(klass, filename, options={}, &block)
     file = File.join(CSV_PATH, filename)
-    CSV.open(file, "r", {:headers => :first_row, :encoding => "utf-8"}) do |csv|
-      klass.destroy_all
-      csv.each do |row|
-        if block_given?
-          yield row
-        else
-          klass.new(row.to_hash).save!
+    if File.exists?(file)
+      CSV.open(file, "r", {:headers => :first_row, :encoding => "utf-8"}) do |csv|
+        klass.destroy_all
+        csv.each do |row|
+          if block_given?
+            yield row
+          else
+            klass.new(row.to_hash).save!
+          end
         end
       end
+      puts " #{Time.now} #{file} -> #{klass.table_name}"
     end
-    puts " #{Time.now} #{file} -> #{klass.table_name}"
   end
 
 end
